+++ .eslintrc.js
@@ -14,7 +14,13 @@ module.exports = {
   ignorePatterns: '*.js',
   overrides: [
     {
-      files: ['packages/*/src/**/*.ts','packages/*/test/**/*.ts', 'src/**/*.ts', 'test/**/*.ts'],
+      files: [
+          'packages/*/src/**/*.ts',
+          'packages/*/test/**/*.ts',
+          'src/**/*.ts',
+          'test/**/*.ts',
+          'examples/**/*.ts',
+      ],
       rules: {
         'indent': ['error', 'tab'],
         'linebreak-style': ['error', 'unix'],